refactor(wallet): extract Client::NetworkOps and Client::AuthenticationOps (#579)
